Move thread-related components
