+++ client/views/room/contextualBar/Threads/ThreadList.tsx
@@ -3,7 +3,6 @@ import { useResizeObserver, useMutableCallback, useAutoFocus } from '@rocket.cha
 import React, { FC, useMemo } from 'react';
 import { Virtuoso } from 'react-virtuoso';
 
-import ThreadComponent from '../../../../../app/threads/client/components/ThreadComponent';
 import { IMessage } from '../../../../../definition/IMessage';
 import { IRoom } from '../../../../../definition/IRoom';
 import { IUser } from '../../../../../definition/IUser';
@@ -17,6 +16,7 @@ import {
 import { useSetting } from '../../../../contexts/SettingsContext';
 import { useTranslation } from '../../../../contexts/TranslationContext';
 import { useTabContext } from '../../providers/ToolboxProvider';
+import ThreadComponent from '../../threads/ThreadComponent';
 import Row from './Row';
 import { withData } from './withData';
